Refactored the code according to the WLF best practice
