+++ webviz_subsurface/plugins/_well_analysis/_utils/__init__.py
@@ -1,0 +1,1 @@
+from ._ensemble_well_analysis_data import EnsembleWellAnalysisData
